requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
gunicorn==22.0.0
Markdown==3.7
packaging==24.0
psycopg2-binary==2.9.9
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
